project2/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pydantic-settings==2.7.1
cryptography==44.0.0
python-multipart==0.0.20
httpx==0.28.1
pytest==8.3.4
minio==7.2.12
